// src/lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const AvailabilityStatus = {
  IN: 'IN',
  OUT: 'OUT',
  PENDING: 'PENDING'
} as const

export type AvailabilityStatus = (typeof AvailabilityStatus)[keyof typeof AvailabilityStatus]


export const ShiftType = {
  HR12: 'HR12',
  HR24: 'HR24',
  HR48: 'HR48'
} as const

export type ShiftType = (typeof ShiftType)[keyof typeof ShiftType]


export const Qualifications = {
  EMT: 'EMT',
  C89: 'C89',
  B85: 'B85',
  DPO: 'DPO',
  DAO: 'DAO'
} as const

export type Qualifications = (typeof Qualifications)[keyof typeof Qualifications]


export const FirefighterRank = {
  DOT: 'DOT',
  ROOKIE: 'ROOKIE',
  RELEASED: 'RELEASED',
  SENIOR: 'SENIOR',
  MASTER: 'MASTER',
  LIEUTENANT: 'LIEUTENANT',
  CAPTAIN: 'CAPTAIN'
} as const

export type FirefighterRank = (typeof FirefighterRank)[keyof typeof FirefighterRank]